*** empty log message ***

+++ StRoot/macros/embedding/daqtags/keepgood.sh
@@ -1,5 +1,19 @@
 #!/bin/bash
 
+if [[ ! $HOST =~ "rcas" ]] ; then
+   echo this code can only be used for RCF!
+   exit
+fi
+
+if [[ ! $# -eq 1 ]] ; then
+   echo ""
+   echo " Usage : $0 [daqfilelist.daq.list] "
+   echo " list of good runs must be stored in goodruns.txt!"
+   echo " output filelist : good_daqfilelist.daq.list. "
+   echo ""
+   exit
+fi
+
 rm -f good_$1
 
 for i in `cat $1`
